Move attack protection to dedicated pkg
